requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-jose
python-multipart
tortoise-orm
alembic
pytest
requests
schema